REVERT: Disable problematic code causing app hang on startup
Disabled the following features that were causing the app to freeze:
1. Nuclear reset in SelineApp.init()
2. Calendar sync in didBecomeActiveNotification
3. Location permission requests in MainAppView.onAppear

These changes were too aggressive and caused the app to hang on load.
Disabling them temporarily so we can isolate the real issue.

After rebuild, the app should load normally and be responsive.
Once working, we can re-enable these features one by one to find
which one is actually causing the freeze.